TSYP_CS/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
groq==0.4.0
pymupdf==1.23.8
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
chromadb==0.4.18
PyGithub==2.1.1
pandas==2.1.4